+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.2.0] - Unreleased
 
+### Added
+-   Add a "Get as..." control to download a single concept or scheme as JSON-LD, Turtle, N-Triples, or RDF/XML [#771](https://github.com/archesproject/arches-lingo/pull/771)
+
 ## [1.1.1] - 2026-07-31
 
 ### Fixed